refactor(server): Move ServerRole to its own module
Role has to decide behaviour below the server: an owner must not open a
login window, and that call is made in dependencies, which server reaches
only through the tool modules. Importing back would close the cycle, so
the enum moves to a module with no package imports of its own.

Re-exported from server, so existing importers are unaffected.

See also: #606
