+++ linkedin_mcp_server/server.py
@@ -6,7 +6,6 @@ person profiles, company data, job information, and session management capabilit
 """
 
 import asyncio
-import enum
 import logging
 from typing import Any, AsyncIterator
 
@@ -28,6 +27,7 @@ from linkedin_mcp_server.error_handler import raise_tool_error
 from linkedin_mcp_server.sequential_tool_middleware import (
     SequentialToolExecutionMiddleware,
 )
+from linkedin_mcp_server.server_role import ServerRole
 from linkedin_mcp_server.update_check import UpdateNoticeMiddleware
 from linkedin_mcp_server.tools.company import register_company_tools
 from linkedin_mcp_server.tools.feed import register_feed_tools
@@ -37,39 +37,6 @@ from linkedin_mcp_server.tools.person import register_person_tools
 from linkedin_mcp_server.tools.post import register_post_tools
 
 logger = logging.getLogger(__name__)
-
-
-class ServerRole(enum.Enum):
-    """Which job a server process does for the shared LinkedIn profile.
-
-    One process per MCP client is the transport's doing, not a choice: a stdio
-    server is spawned per client instance. That makes "who drives Chromium" a
-    property of the process rather than of the code, and every difference below
-    follows from it.
-    """
-
-    #: Drives its own browser and talks to its own client. The historical
-    #: behaviour, and still what an explicit HTTP bind or an embedder gets.
-    DIRECT = "direct"
-
-    #: Drives the browser on behalf of other processes over loopback HTTP.
-    #: Never speaks to an end client, so nothing user-facing belongs here.
-    OWNER = "owner"
-
-    # A forwarding role belongs here too, but only once something actually
-    # forwards. Adding it now would mean a server that registers the local
-    # browser-backed tools and then declines to serialize them, which is a
-    # worse starting point than not having the role at all.
-
-    @property
-    def drives_browser(self) -> bool:
-        """Whether this role launches Chromium against the shared profile."""
-        return self in (ServerRole.DIRECT, ServerRole.OWNER)
-
-    @property
-    def faces_a_client(self) -> bool:
-        """Whether an end user reads this server's tool results."""
-        return self is ServerRole.DIRECT
 
 
 @lifespan
